Renamed the model and service to follow camelCase
